requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.50
PyYAML>=6.0.1
PySide6>=6.7.0
pytz>=2024.1
rich>=13.7.1
tqdm>=4.66.4
